refactoring

+++ include/decompress.h
@@ -29,6 +29,6 @@ void HandleLoadSpecialPokePic(bool32 isFrontPic, void *dest, s32 species, u32 pe
 
 void LoadSpecialPokePic(void *dest, s32 species, u32 personality, bool8 isFrontPic);
 
-u32 GetDecompressedDataSize(const u8 *ptr);
+u32 GetDecompressedDataSize(const u32 *ptr);
 
 #endif // GUARD_DECOMPRESS_H
